¡¡¡ "Several" changes for this entire project !!!
Commit's features:

- Added new uniform: Hoshinomiya School Winter Uniform
- Important fixes in the character's skin tone for all the characters on the following folders:
	- Blue Reflection --> Vacation Style
	- Blue Reflection --> Summer Clothes
	- Blue Reflection --> Student Skins
- Removed the Heads folder for the next release: requires fixes based in the new skin tones

Important Note #1: from now, the future changes for each work day on this repository will be documented only on each commit description. The new features, changes, fixes and notes will be described only for the releases' description.

Important Note #2: for the next pack release (v1.1), there'll be some restructuring in the manual, mainly on the ways of how to download the skins pack. Also, the next version of the pack will be focused mainly on fixes and aesthetic changes.